+++ pyviews/binding/tests/inline_tests.py
@@ -6,7 +6,7 @@ from pytest import fixture, mark
 from pyviews.binding import BindingContext, InlineBinding
 from pyviews.binding.inline import bind_inline
 from pyviews.binding.tests.common import InnerViewModel
-from pyviews.expression import Expression
+from pyviews.core.expression import Expression
 from pyviews.core import InheritedDict, XmlAttr, BindingCallback